feat(core)!: remove the class-based dialog UI and go headless
BREAKING CHANGE: the imperative dialog layer is gone. EditorDialog,
ItemEditorDialog, BlockCatalogDialog, BlockOptionsDialog, BlockMenu,
InitialInsertionButton, ItemEditorService and getItemEditorTemplate
are removed, along with the UIOptions/UICreator/EditorDialogShellCreator
DI contracts and the [data-bge-editor-ui] marker scanning. The UI layer
now subscribes to engine.uiState and renders dialogs declaratively.

- Item no longer holds an editor dialog or a click listener; opening
  the editor is a UI-layer concern (uiState.openItemEditor). Item
  exposes its seed; import() is synchronous and no longer runs
  beforeChange
- ItemEditorOptions keeps only isDisable; the editor string field,
  open/beforeOpen/beforeChange/onSubmit hooks and customData are gone
  (replaced by Editor/toEditorState/toItemData)
- applyBlockOptions(): the block options form application extracted
  from BlockOptionsDialog for reuse by the declarative dialog
- engine exposes catalog and getContentStylesheet(); EditableArea
  installs a command bus receiver in its iframe document and the
  fallback initial-insertion button declares an invoker command
  instead of a click listener
- Actions loses open-editor / select-tab-in-item-editor (editor-tree
  coordination is lifted into React state)
