+++ packages/@burger-editor/core/src/engine/engine.ts
@@ -1,9 +1,8 @@
 import type { ContainerType } from '../block/types.js';
-import type { DialogSettings } from '../editor-dialog.js';
 import type { ItemSeed } from '../item/types.js';
 import type {
 	BurgerEditorEngineOptions,
-	UIOptions,
+	BlockCatalog,
 	Config,
 	FileAPI,
 	BlockItem,
@@ -11,8 +10,6 @@ import type {
 } from '../types.js';
 
 import { BurgerBlock } from '../block/block.js';
-import { BlockCatalogDialog } from '../block-catalog-dialog.js';
-import { BlockOptionsDialog } from '../block-options-dialog.js';
 import { CommandBus } from '../command/command-bus.js';
 import { ComponentObserver } from '../component-observer.js';
 import { CSS_LAYER } from '../const.js';
@@ -28,17 +25,14 @@ import { getElement } from '../dom-helpers/get-element.js';
 import { EditableArea } from '../editable-area.js';
 import { createBgeEvent } from '../event/create-bge-event.js';
 import { HealthMonitor } from '../health-monitor.js';
-import { getItemEditorTemplate } from '../item/get-item-editor-template.js';
 import { Item } from '../item/item.js';
-import { ItemEditorDialog } from '../item-editor-dialog.js';
 
 import { UIStateStore } from './ui-state.js';
 
 type ConfirmCallback = () => Promise<boolean> | boolean;
 
 export class BurgerEditorEngine {
-	readonly blockCatalogDialog: BlockCatalogDialog;
-	readonly blockOptionsDialog: BlockOptionsDialog;
+	readonly catalog: BlockCatalog;
 	readonly commandBus = new CommandBus();
 	readonly componentObserver = new ComponentObserver();
 	readonly config: Config;
@@ -51,13 +45,11 @@ export class BurgerEditorEngine {
 		readonly generalCSS: string;
 	};
 	readonly el: HTMLElement;
-	readonly itemEditorDialog: ItemEditorDialog<{}, {}>;
 	readonly items: Map<string, ItemSeed>;
 	readonly serverAPI: FileAPI;
 	readonly storageKey: {
 		readonly blockClipboard: string;
 	};
-	readonly ui: UIOptions;
 	readonly uiState = new UIStateStore();
 	readonly viewArea: HTMLElement;
 	#contentStylesheetCache: string | null = null;
@@ -87,7 +79,7 @@ export class BurgerEditorEngine {
 		this.el = getElement(options.root);
 		this.config = options.config;
 		this.serverAPI = options.fileIO ?? {};
-		this.ui = options.ui ?? {};
+		this.catalog = options.catalog;
 		this.storageKey = {
 			blockClipboard: 'bge-copied-block',
 			...options.storageKey,
@@ -120,78 +112,6 @@ export class BurgerEditorEngine {
 			script.src = `https://maps.googleapis.com/maps/api/js?key=${this.config.googleMapsApiKey}&libraries=marker`;
 			document.head.append(script);
 		}
-
-		const dialogSettings: DialogSettings = {
-			onClosed: () => {
-				this.componentObserver.off();
-				this.save();
-			},
-			onOpen: () => {
-				return this.isProcessed;
-			},
-			createEditorComponent: (el) => {
-				const editorComponentSubClassName = el.dataset.bgeEditorUi;
-				if (editorComponentSubClassName && this.#isUIName(editorComponentSubClassName)) {
-					const cleanUpHook = this.ui[editorComponentSubClassName]?.(el, this);
-					return cleanUpHook?.cleanUp;
-				}
-				return;
-			},
-			createDialogShell: options.dialogShell,
-		};
-
-		this.blockCatalogDialog = new BlockCatalogDialog(options.catalog, {
-			...dialogSettings,
-			addBlock: (blockData) => {
-				return this.addBlock(blockData);
-			},
-		});
-
-		this.blockOptionsDialog = new BlockOptionsDialog({
-			...dialogSettings,
-			onChangeBlock: (callback) => {
-				this.el.addEventListener('bge:block-change', (e) => {
-					callback(e.detail.block);
-				});
-			},
-			getCurrentBlock: () => {
-				return this.getCurrentBlock();
-			},
-		});
-
-		this.itemEditorDialog = new ItemEditorDialog({
-			...dialogSettings,
-			config: this.config,
-			onOpened: (data, editor) => {
-				this.componentObserver.notify('open-editor', {
-					data,
-					editor,
-				});
-			},
-			getComponentObserver: () => {
-				return this.componentObserver;
-			},
-			getTemplate: (itemName: string) => {
-				return getItemEditorTemplate(this, itemName);
-			},
-			getContentStylesheet: async () => {
-				if (this.#contentStylesheetCache) {
-					return this.#contentStylesheetCache;
-				}
-				const css = await Promise.all(
-					this.css.stylesheets
-						.filter((sheet) => sheet.layer == null)
-						.map(async (sheet) => {
-							const res = await fetch(sheet.path);
-							return res.text();
-						}),
-				);
-				// generalCSSを含める
-				const stylesheets = [this.css.generalCSS, ...css];
-				this.#contentStylesheetCache = stylesheets.join('\n');
-				return this.#contentStylesheetCache;
-			},
-		});
 
 		this.items = new Map();
 		if (options.items) {
@@ -257,6 +177,28 @@ export class BurgerEditorEngine {
 		return false;
 	}
 
+	/**
+	 * Resolve the CSS applied to the content (generalCSS plus non-layered
+	 * stylesheets), for injection into rich-text editors.
+	 * @returns The concatenated stylesheet text
+	 */
+	async getContentStylesheet(): Promise<string> {
+		if (this.#contentStylesheetCache) {
+			return this.#contentStylesheetCache;
+		}
+		const css = await Promise.all(
+			this.css.stylesheets
+				.filter((sheet) => sheet.layer == null)
+				.map(async (sheet) => {
+					const res = await fetch(sheet.path);
+					return res.text();
+				}),
+		);
+		// generalCSSを含める
+		const stylesheets = [this.css.generalCSS, ...css];
+		this.#contentStylesheetCache = stylesheets.join('\n');
+		return this.#contentStylesheetCache;
+	}
 	getCurrentBlock() {
 		if (!this.#currentBlock) {
 			// eslint-disable-next-line no-console
@@ -381,7 +323,7 @@ export class BurgerEditorEngine {
 
 	async #createItemElement(itemData: BlockItem | HTMLElement) {
 		if (typeof itemData !== 'string' && 'localName' in itemData) {
-			const item = Item.rebind(itemData, this.items, this.itemEditorDialog);
+			const item = Item.rebind(itemData, this.items);
 			return item.el;
 		}
 
@@ -389,14 +331,10 @@ export class BurgerEditorEngine {
 		const item = await Item.create(
 			name,
 			this.items,
-			this.itemEditorDialog,
+			this.config,
 			typeof itemData === 'string' ? undefined : itemData.data,
 		);
 		return item.el;
-	}
-
-	#isUIName(name: string): name is keyof UIOptions {
-		return name in this.ui;
 	}
 
 	/**
