+++ packages/@burger-editor/core/src/index.ts
@@ -7,7 +7,7 @@ export { BurgerBlock } from './block/block.js';
 export { BurgerEditorEngine } from './engine/engine.js';
 export { Item } from './item/item.js';
 export { createItem } from './item/create-item.js';
-export { ItemEditorDialog } from './item-editor-dialog.js';
+export { applyBlockOptions } from './block/apply-block-options.js';
 export { dataToHtml as itemImport } from './item/data-to-html.js';
 export { dataFromHtml as itemExport } from './item/data-from-html.js';
 export { render } from './render.js';
