+++ packages/@burger-editor/core/src/item/item.ts
@@ -1,4 +1,3 @@
-import type { ItemEditorDialog } from '../item-editor-dialog.js';
 import type { ItemData, ItemSeed } from './types.js';
 import type { Config } from '../types.js';
 
@@ -7,7 +6,6 @@ import { replacePlaceholders } from '../utils/replace-placeholders.js';
 import { createUnknownContentItem } from './create-item.js';
 import { dataFromHtml } from './data-from-html.js';
 import { dataToHtml } from './data-to-html.js';
-import { ItemEditorService } from './item-editor-service.js';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const elMap = new WeakMap<HTMLElement, Item<any, any>>();
@@ -17,10 +15,9 @@ export class Item<
 	C extends { [key: string]: unknown },
 	N extends keyof T & string = keyof T & string,
 > {
-	readonly editor: ItemEditorDialog<T, C>;
 	readonly name: string;
+	readonly seed: ItemSeed<N, T, C>;
 	readonly #el: HTMLElement;
-	readonly #service: ItemEditorService<T, C, N>;
 	#version: string;
 
 	get el() {
@@ -32,65 +29,46 @@ export class Item<
 	}
 
 	// eslint-disable-next-line no-restricted-syntax
-	private constructor(
-		seed: ItemSeed<N, T, C> | null,
-		el: HTMLElement,
-		editor: ItemEditorDialog<T, C>,
-	) {
+	private constructor(seed: ItemSeed<N, T, C> | null, el: HTMLElement) {
 		elMap.set(el, this);
 		this.#el = el;
-
-		this.#el.addEventListener('click', (e) => {
-			e.preventDefault();
-			e.stopPropagation();
-			void this.#openEditor();
-		});
 
 		// Synthesize fallback seed when missing
 		const effectiveSeed = seed ?? createUnknownContentItem<T, C, N>(el);
 
 		this.name = effectiveSeed.name;
 		this.#version = effectiveSeed.version;
-
-		this.#service = new ItemEditorService<T, C, N>(this, effectiveSeed);
-		this.editor = editor;
+		this.seed = effectiveSeed;
 	}
 
 	export() {
 		return dataFromHtml(this.el.innerHTML) as T;
 	}
 
-	async import(newData: Partial<T>, fromDialogChanges = false) {
+	import(newData: Partial<T>) {
 		const currentData = this.export();
-		let data: T = {
+		const data: T = {
 			...currentData,
 			...newData,
 		};
-		if (fromDialogChanges) {
-			data = await this.#service.beforeChange(data, this.editor);
-		}
 
 		this.el.innerHTML = Item.createElement(this.el.innerHTML, data);
 	}
 
 	isDisable() {
-		return this.#service.isDisable(this);
-	}
-
-	async #openEditor() {
-		await this.editor.open(this.#service);
+		return this.seed.editorOptions?.isDisable?.(this) ?? '';
 	}
 
 	static async create<T extends ItemData, C extends { [key: string]: unknown }>(
 		name: string,
 		// eslint-disable-next-line @typescript-eslint/no-explicit-any
 		itemSeeds: ReadonlyMap<string, ItemSeed<string, any, any>>,
-		editor: ItemEditorDialog<T, C>,
+		config: Config,
 		initData: Partial<T> = {},
 	) {
 		const seed: ItemSeed<string, T, C> | null = itemSeeds.get(name) ?? null;
-		const wrapper = Item.createWrapper(name, seed, editor.config);
-		const item = new Item<T, C>(seed, wrapper, editor);
+		const wrapper = Item.createWrapper(name, seed, config);
+		const item = new Item<T, C>(seed, wrapper);
 		await item.import(initData);
 		return item;
 	}
@@ -120,14 +98,13 @@ export class Item<
 		el: HTMLElement,
 		// eslint-disable-next-line @typescript-eslint/no-explicit-any
 		itemSeeds: ReadonlyMap<string, ItemSeed<string, any, any>>,
-		editor: ItemEditorDialog<T, C>,
 	) {
 		const name = el.dataset.bgi;
 		if (!name) {
 			throw new Error('data-bgi not found');
 		}
 		const seed: ItemSeed<string, T, C> | null = itemSeeds.get(name) ?? null;
-		const item = new Item<T, C>(seed, el, editor);
+		const item = new Item<T, C>(seed, el);
 		return item;
 	}
